Retinopathy_Back/requirements.txt
asgiref==3.8.1
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
flatbuffers==24.3.25
fonttools==4.51.0
humanfriendly==10.0
kiwisolver==1.4.5
matplotlib==3.8.4
mpmath==1.3.0
numpy==1.26.4
onnxruntime==1.17.3
packaging==24.0
pillow==10.3.0
protobuf==5.26.1
pyparsing==3.1.2
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
sympy==1.12
typing_extensions==4.11.0
tzdata==2024.1